Adapted to Poetry for packaging and dependency management
