+++ pyproject.toml
@@ -11,10 +11,16 @@ packages = [
 
 [tool.poetry.dependencies]
 python = "^3.6"
-tensorflow = "^2.1"
+typer = "^0.1.1"
+numpy = "^1.18.3"
 
 [tool.poetry.dev-dependencies]
+tensorflow = "^2.1.0"
 pytest = "^5.4.1"
+tensorflow-estimator = "2.1"
+
+[tool.poetry.scripts]
+konverter = 'konverter.__main__:run'
 
 [build-system]
 requires = ["poetry>=0.12"]
